feat: create Pages directory with Login, Signup, Dashboard, and Home without components
