+++ frontend/src/App.js
@@ -1,29 +1,17 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-
-function Home() {
-  return <h1>Home Page</h1>;
-}
-
-function Signup() {
-  return <h1>Signup Page</h1>;
-}
-
-function Login() {
-  return <h1>Login Page</h1>;
-}
-
-function Dashboard() {
-  return <h1>Dashboard Page</h1>;
-}
+import Home from "./pages/Home";
+import Login from "./pages/Login";
+import Signup from "./pages/Signup";
+import Dashboard from "./pages/Dashboard";
 
 function App() {
   return (
     <Router>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/signup" element={<Signup />} />
         <Route path="/login" element={<Login />} />
+        <Route path="/signup" element={<Signup />} />
         <Route path="/dashboard" element={<Dashboard />} />
       </Routes>
     </Router>
