+++ frontend/src/pages/Dashboard.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Dashboard() {
+  return <h1>Dashboard Page</h1>;
+}
+
+export default Dashboard;
