+++ frontend/src/pages/Home.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Home() {
+  return <h1>Home Page</h1>;
+}
+
+export default Home;
