+++ frontend/src/pages/Signup.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Signup() {
+  return <h1>Signup Page</h1>;
+}
+
+export default Signup;
